Set up simplified skeleton of the evaluator.
For now, this is a simple, conventional meta-circular evaluator
of the sort described in innumerable Scheme and Lisp textbooks,
though at the moment it only can recognize and 'evaluate' simple
literals. This is just meant to test the basic behavior of the
elan:eval procedure. For now, the process will proceed by
filling out the usual details of the evaluator. Once the typical
basic functionality of Scheme is working, this will be refactored
into a testbed for the enclosed-table approach that is the primary
testing goal of the project.